requirements.txt
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.7.7
googlemaps==4.10.0 